feat: fix LoanPosition.mint properly

+++ contracts/interfaces/ILoanPositionManager.sol
@@ -53,7 +53,7 @@ interface ILoanPositionManager is IERC721Metadata, IERC721Permit, ILoanPositionM
         address debtToken,
         uint256 collateralAmount,
         uint256 debtAmount,
-        uint256 loanEpochs,
+        uint16 loanEpochs,
         address recipient,
         bytes calldata data
     ) external returns (uint256);
